parkinson_voice_model/src/requirements.txt
# Core ML and Data Processing Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Audio Processing Libraries  
librosa>=0.8.1
praat-parselmouth>=0.4.0
soundfile>=0.10.0
scipy>=1.7.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Model Persistence and Utilities
joblib>=1.1.0
openpyxl>=3.0.0

# Progress Tracking
tqdm>=4.62.0

# Optional: For SHAP interpretability
shap>=0.40.0
